Added keywords and messages to the database and interface
We now have a functional suite execution. Two collections are used,
one for suites (that holds suites, tests and keywords, so we may
need a better name) and one for messages.

Message ordering is not considered yet. For instance, there might
be messages issued before and after children keywords were called.
We might consider the order we receive messages here.